requirements.txt
pymodbus
pyserial
retry